Workaround for some unit test issues
